+++ src/templates/artifacts/SimpleJpaModelChild.groovy
@@ -11,27 +11,19 @@ import simplejpa.swing.*
 class $className {
 
     @Bindable Long id
-<% fields.each { field ->
-    if (["BASIC_TYPE", "DATE"].contains(field.info)) {
-        out << "\t@Bindable ${field.type} ${field.name}\n"
-    } else if ("UNKNOWN".equals(field.info)){
-        out << "\t// ${field.name} is not supported by generator.  You will need to code it manually.\n"
-        out << "\t@Bindable ${field.type} ${field.name}\n"
-    }
-} %>
-    @Bindable String ${fields[0]?.name ?: 'replaceThis'}Search
-    @Bindable String searchMessage
-<%  fields.each { field ->
-        if (field.info=="DOMAIN_CLASS" && !field.annotations?.containsAttribute('mappedBy')) {
+<%
+    fields.each { field ->
+        if (["BASIC_TYPE", "DATE"].contains(field.info)) {
+            out << "\t@Bindable ${field.type} ${field.name}\n"
+        } else if ("UNKNOWN".equals(field.info)){
+            out << "\t// ${field.name} is not supported by generator.  You will need to code it manually.\n"
+            out << "\t@Bindable ${field.type} ${field.name}\n"
+        } else if (field.info=="DOMAIN_CLASS" && !field.annotations?.containsAttribute('mappedBy')) {
             out << "\tBasicEventList<${field.type}> ${field.name}List = new BasicEventList<>()\n"
             out << "\t@Bindable DefaultEventComboBoxModel<${field.type}> ${field.name} =\n"
             out << "\t\tGlazedListsSwing.eventComboBoxModelWithThreadProxyList(${field.name}List)\n"
-        } else if (field.type.toString()=="List" && field.info!="UNKNOWN") {
-            if (field.annotations?.get("OneToMany")!=null) {
-                out << "\tList<${field.info}> ${field.name} = []\n"
-            } else {
-                out << "\tTagChooserModel ${field.name} = new TagChooserModel()\n"
-            }
+        } else if (field.type.toString()=="List" && field.info!="UNKNOWN" && field.annotations?.get("OneToMany")==null) {
+            out << "\tTagChooserModel ${field.name} = new TagChooserModel()\n"
         }
     }
 %>
@@ -54,13 +46,8 @@ class $className {
             out << "\t\t\t\t${field.name} = selected.${field.name}\n"
         } else if (field.info=="DOMAIN_CLASS" && !field.annotations?.containsAttribute('mappedBy')) {
             out << "\t\t\t\t${field.name}.selectedItem = selected.${field.name}\n"
-        } else if (field.type.toString()=="List" && field.info!="UNKNOWN") {
-            if (field.annotations?.get("OneToMany")!=null) {
-                out << "\t\t\t\t${field.name}.clear()\n"
-                out << "\t\t\t\t${field.name}.addAll(selected.${field.name})\n"
-            } else {
-                out << "\t\t\t\t${field.name}.replaceSelectedValues(selected.${field.name})\n"
-            }
+        } else if (field.type.toString()=="List" && field.info!="UNKNOWN" && field.annotations?.get("OneToMany")==null) {
+            out << "\t\t\t\t${field.name}.replaceSelectedValues(selected.${field.name})\n"
         } else if (field.info=="UNKNOWN") {
             out << "\t\t\t\t// ${field.name} is not supported by generator.  You will need to code it manually.\n"
             out << "\t\t\t\t${field.name} = selected.${field.name}\n"
@@ -73,17 +60,13 @@ class $className {
 
     def clear = {
         id = null
-<% fields.collect { field ->
+<% fields.each { field ->
         if (["BASIC_TYPE", "DATE"].contains(field.info)) {
             out << "\t\t${field.name} = null\n"
         } else if (field.info=="DOMAIN_CLASS" && !field.annotations?.containsAttribute('mappedBy')) {
             out << "\t\t${field.name}.selectedItem = null\n"
-        } else if (field.type.toString()=="List" && field.info!="UNKNOWN") {
-            if (field.annotations?.get("OneToMany")!=null) {
-                out << "\t\t${field.name}.clear()\n"
-            } else {
-                out << "\t\t${field.name}.clearSelectedValues()\n"
-            }
+        } else if (field.type.toString()=="List" && field.info!="UNKNOWN" && field.annotations?.get("OneToMany")==null) {
+            out << "\t\t${field.name}.clearSelectedValues()\n"
         } else if (field.info=="UNKNOWN") {
             out << "\t\t// ${field.name} is not supported by generator.  You will need to code it manually.\n"
             out << "\t\t${field.name} = null\n"